requirements.txt
alembic==1.13.1
anyio==4.4.0
certifi==2024.6.2
dnspython==2.6.1
exceptiongroup==1.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pymongo==4.7.2
python-telegram-bot==21.2
pytz==2024.1
sniffio==1.3.1
SQLAlchemy==2.0.30
typing_extensions==4.12.1
zipp==3.19.1
